made fetch note api into a get request
